Fix CUDA error message formatting to cast error code to int

+++ src/scalable_ccd/cuda/utils/assert.cuh
@@ -21,7 +21,8 @@ inline void gpuAssert(
             cudaGetErrorString(code), file, line);
         if (throw_on_error) {
             assert(code == cudaSuccess);
-            throw std::runtime_error(fmt::format("CUDA error {}", code));
+            throw std::runtime_error(
+                fmt::format("CUDA error {}", static_cast<int>(code)));
         }
     }
 }
